+++ src/main/java/com/uor/eng/Main.java
@@ -6,6 +6,7 @@ import javafx.stage.Stage;
 /**
  * The main JavaFX Application entry point.
  */
+
 public class Main extends Application {
   private final MonitoringController controller;
